revert(analytics): drop feature-usage endpoint discount support
feature_usage_tracking.go is being deprecated in favor of the meter_usage
pipeline, so percentage-discount support there isn't worth carrying. Fully
reverted to its pre-PR state (the applyAnalyticsDiscounts wiring in
buildAnalyticsResponse, the DTO field mapping, and the aggregateAnalyticsByGrouping/
mergeTimeSeriesPoints discount-carry-forward fix), and removed the two test
files written for it (feature_usage_discount_test.go,
feature_usage_grouping_discount_test.go).

Also applies a simplification flagged by code review (I1): AnalyticsData's
LineItemCoupons/SubscriptionCoupons fields were only ever read inside
applyAnalyticsDiscounts itself — no other caller in either analytics path
touched them. Now that the feature-usage path no longer needs to share
AnalyticsData's shape for this, they're local variables in
applyAnalyticsDiscounts instead of struct fields on a type shared across the
whole analytics pipeline.

Also relocates TestApplyAnalyticsDiscounts_EmptyInputsAreNoOp (a pure
applyAnalyticsDiscounts unit test with no dependency on the
MeterUsageDiscountSuite fixture) from meter_usage_discount_test.go to
analytics_discount_test.go, alongside its sibling pure-function tests (review
finding M2).

Meter-usage detailed-analytics discount support (the primary target) is
unaffected — all tests still pass, coverage unchanged.
